feat: update specialty routes and service to use patch method and improve error handling

+++ server/src/app/module/specialty/specialty.service.ts
@@ -1,5 +1,7 @@
+import status from "http-status";
 import { Specialty } from "../../../generated/prisma/client";
 import { prisma } from "../../libs/prisma";
+import AppError from "../../shared/appError";
 
 const createSpecialty = async (payload: Specialty): Promise<Specialty> => {
   const specialty = await prisma.specialty.create({
@@ -17,12 +19,19 @@ const getSpecialtyById = async (id: string): Promise<Specialty | null> => {
       id,
     },
   });
+  if (!specialty) {
+    throw new AppError(status.NOT_FOUND, "Specialty not found");
+  }
   return specialty;
 };
 const updateSpecialty = async (
   id: string,
   payload: Specialty,
 ): Promise<Specialty> => {
+  const isExist = await prisma.specialty.findUnique({ where: { id } });
+  if (!isExist) {
+    throw new AppError(status.NOT_FOUND, "Specialty is not found");
+  }
   const specialty = await prisma.specialty.update({
     where: {
       id,
@@ -32,6 +41,10 @@ const updateSpecialty = async (
   return specialty;
 };
 const deleteSpecialty = async (id: string): Promise<Specialty> => {
+  const isExist = await prisma.specialty.findUnique({ where: { id } });
+  if (!isExist) {
+    throw new AppError(status.NOT_FOUND, "Specialty is not found");
+  }
   const specialty = await prisma.specialty.delete({
     where: {
       id,
